fix(dossier): decoralate dropdown behaviour from autosave
