+++ app/models/champ.rb
@@ -49,6 +49,9 @@ class Champ < ApplicationRecord
     :dossier_link?,
     :titre_identite?,
     :header_section?,
+    :simple_drop_down_list?,
+    :linked_drop_down_list?,
+    :non_fillable?,
     :cnaf?,
     :dgfip?,
     :pole_emploi?,
